Move execution context to application layer
