+++ model_service/application/command_callbacks/core/execution_context/create_execution_context.py
@@ -1,7 +1,7 @@
 from pika import BasicProperties
 
 from model_service.application.command_callbacks.core.execution_context.exceptions import ExecutionContextCreationError
-from model_service.domain.shared.execution_context import ExecutionContext
+from model_service.application.shared.execution_context import ExecutionContext
 
 
 def create_execution_context(properties: BasicProperties) -> ExecutionContext:
