+++ model_service/application/schema_mappers/map_feed_forward_network_from_schema.py
@@ -2,7 +2,7 @@ from model_service.application.schemas.feed_forward_network_schema import FeedFo
 from model_service.domain.entities.neural_network.feed_forward_network import FeedForwardNetwork
 from model_service.domain.entities.value_objects.activation_function import ActivationFunction, ActivationFunctionType
 from model_service.domain.entities.value_objects.linear_layer import LinearLayer
-from model_service.domain.shared.execution_context import ExecutionContext
+from model_service.application.shared.execution_context import ExecutionContext
 from model_service.shared.dependency_management.dependency_kind import SCOPED
 from model_service.shared.dependency_management.provide import provide, Dependency
