+++ model_service/application/schema_mappers/map_training_session_from_schema.py
@@ -3,7 +3,7 @@ from model_service.domain.entities.training.loss_function_type import LossFuncti
 from model_service.domain.entities.training.metric_type import MetricType
 from model_service.domain.entities.training.training_session import TrainingSession
 from model_service.domain.entities.value_objects.optimizer import Optimizer
-from model_service.domain.shared.execution_context import ExecutionContext
+from model_service.application.shared.execution_context import ExecutionContext
 from model_service.shared.dependency_management.dependency_kind import SCOPED
 from model_service.shared.dependency_management.provide import Dependency, provide
