+++ model_service/application/shared/execution_context.py
