+++ tests/test_start_ffn_training.py
@@ -20,7 +20,7 @@ from model_service.domain.repositories.abstract_feed_forward_network_repository 
 from model_service.domain.repositories.abstract_table_dataset_repository import AbstractTableDatasetRepository
 from model_service.domain.repositories.abstract_training_result_repository import AbstractTrainingResultRepository
 from model_service.domain.repositories.abstract_training_session_repository import AbstractTrainingSessionRepository
-from model_service.domain.shared.execution_context import ExecutionContext
+from model_service.application.shared.execution_context import ExecutionContext
 from model_service.shared.dependency_management.provider_getters import get_scoped_provider
 from model_service.shared.dependency_management.providers.scoped_dependency_provider import ScopedDependencies
